Added Dagger for DI and made big refactoring of project
