+++ app/src/main/java/com/example/anatomyviewer/ar/di/ArComponent.kt
@@ -1,0 +1,10 @@
+package com.example.anatomyviewer.ar.di
+import com.example.anatomyviewer.ar.AnatomyViewerFragment
+import dagger.Component
+import javax.inject.Singleton
+
+@Singleton
+@Component(modules = [ArModule::class])
+interface ArComponent {
+    fun inject(target: AnatomyViewerFragment)
+}
