+++ app/src/main/java/com/example/anatomyviewer/ar/quiz/Question.kt
@@ -1,0 +1,9 @@
+package com.example.anatomyviewer.ar.quiz
+
+data class Question(
+    val questionText: String,
+    val opt1Text: String,
+    val opt2Text: String,
+    val opt3Text: String,
+    val correctOption: Int
+) {}
