+++ app/src/main/res/layout/view_quiz_card.xml
@@ -5,8 +5,8 @@
 
     <data>
         <variable
-            name="viewModel"
-            type="com.example.anatomyviewer.ARViewModel" />
+            name="quizData"
+            type="com.example.anatomyviewer.ar.quiz.QuizData" />
     </data>
 
     <RelativeLayout
@@ -20,7 +20,7 @@
             android:id="@+id/text_view_score"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@{viewModel.quizData.scoreLabelText}"
+            android:text="@{quizData.scoreLabelText}"
             android:textColor="@android:color/black" />
 
         <TextView
@@ -28,7 +28,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@id/text_view_score"
-            android:text="@{viewModel.quizData.questionNumbText}"
+            android:text="@{quizData.questionNumbText}"
             android:textColor="@android:color/black" />
 
         <TextView
@@ -36,7 +36,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentEnd="true"
-            android:text="@{viewModel.quizData.timeText}"
+            android:text="@{quizData.timeText}"
             android:textColor="@android:color/black"
             android:textSize="40sp" />
 
@@ -46,7 +46,7 @@
             android:layout_height="wrap_content"
             android:layout_above="@id/radio_group"
             android:layout_marginBottom="16dp"
-            android:text="@{viewModel.quizData.questionText}"
+            android:text="@{quizData.questionText}"
             android:textAlignment="center"
             android:textColor="@android:color/black"
             android:textSize="20sp" />
@@ -65,24 +65,24 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:checked="false"
-                android:onClick="@{() -> viewModel.quizData.opt1Clicked()}"
-                android:text="@{viewModel.quizData.opt1Text}" />
+                android:onClick="@{() -> quizData.opt1Clicked()}"
+                android:text="@{quizData.opt1Text}" />
 
             <RadioButton
                 android:id="@+id/radio_button2"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:checked="false"
-                android:onClick="@{() -> viewModel.quizData.opt2Clicked()}"
-                android:text="@{viewModel.quizData.opt2Text}" />
+                android:onClick="@{() -> quizData.opt2Clicked()}"
+                android:text="@{quizData.opt2Text}" />
 
             <RadioButton
                 android:id="@+id/radio_button3"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:checked="false"
-                android:onClick="@{() -> viewModel.quizData.opt3Clicked()}"
-                android:text="@{viewModel.quizData.opt3Text}" />
+                android:onClick="@{() -> quizData.opt3Clicked()}"
+                android:text="@{quizData.opt3Text}" />
 
         </RadioGroup>
 
@@ -92,7 +92,7 @@
             android:layout_height="wrap_content"
             android:layout_below="@id/radio_group"
             android:layout_marginTop="16dp"
-            android:onClick="@{() -> viewModel.quizData.clickedConfirmed()}"
+            android:onClick="@{() -> quizData.clickedConfirmed()}"
             android:text="Confirm" />
 
     </RelativeLayout>
